Ussless code removed
Acording to \PHPUnit_Util_Configuration::getTestSuite method each  of suite nodes:
* directory
* file
* exclude

is processed by \PHPUnit_Util_Configuration::toAbsolutePath
which returns absolute path if that was passed to it or searches for path relative to directory with configuration file. (XmlConfiguration::makeAbsolutePath)

Thats why adding new nodes and removing old once is not needed.
Only update of nodes is required.
And path is allways resoloved via directory of configuration file, so getcwd() was a bug.